+++ player.cpp
@@ -2,7 +2,7 @@
 
 std::string Player::getName() { return name; }
 void Player::setName(std::string name) { Player::name = name; }
-short Player::getPosition() { return position; }
-void Player::setPosition(short pos) { position = pos; }
-void Player::setTurnsLeftInJail(short turns) {turnsLeftInJail = turns; }
-short Player::getTurnsLeftInJail() { return turnsLeftInJail; }
+int Player::getPosition() { return position; }
+void Player::setPosition(int pos) { position = pos; }
+void Player::setTurnsLeftInJail(int turns) {turnsLeftInJail = turns; }
+int Player::getTurnsLeftInJail() { return turnsLeftInJail; }
